Fixed a Regression in FSD Handling Backspace
- Because Backspace isn't added to the InputBuffer and handled in `ProcessIncomingCharacter`, we have to detect if it's the case in FSD and handle it there as well
- Fixed Typo
- Removed redundant cast in CPU Core

+++ MBBSEmu/CPU/CPUCore.cs
@@ -1126,7 +1126,7 @@ namespace MBBSEmu.CPU
         [MethodImpl(OpcodeCompilerOptimizations)]
         private void Op_Cwd()
         {
-            Registers.DX = Registers.AX.IsBitSet(15) ? (ushort)0xFFFF : (ushort)0x0000;
+            Registers.DX = Registers.AX.IsBitSet(15) ? 0xFFFF : 0x0000;
         }
